+++ Changelog.md
@@ -67,6 +67,7 @@
 - Fix: Stonecutting some types of reinforced copper blocks yields too few blocks compared to vanilla
 - Fix: Reinforced Chiseled Deepslate cannot be crafted with Reinforced Cobbled Deepslate Slabs
 - Removed: "Is always active" option of Keypads and Key Panels. The same behavior can be achieved by setting the "Signal length" option to 0 and activating the block
+- Misc.: Small optimizations to certain network and client performance
 
 --------------------------Changelog for v1.9.8 of SecurityCraft--------------------------